removes make_result API
