+++ autogalaxy/analysis/analysis/dataset.py
@@ -159,8 +159,8 @@ class AnalysisDataset(Analysis):
         Parameters
         ----------
         paths
-            The PyAutoFit paths object which manages all paths, e.g. where the non-linear search outputs are stored, visualization,
-            and the pickled objects used by the aggregator output by this function.
+            The PyAutoFit paths object which manages all paths, e.g. where the non-linear search outputs are stored,
+            visualization, and the pickled objects used by the aggregator output by this function.
         """
         paths.save_fits(
             name="data",
