+++ autogalaxy/analysis/result.py
@@ -4,9 +4,6 @@ from typing import Dict, List, Optional, Tuple, Type, Union
 
 from typing import TYPE_CHECKING
 
-if TYPE_CHECKING:
-    from autogalaxy.analysis.analysis.dataset import AnalysisDataset
-
 from autoconf import cached_property
 
 import autofit as af
@@ -17,54 +14,6 @@ from autogalaxy.galaxy.galaxy import Galaxy
 
 
 class Result(af.Result):
-    def __init__(
-        self,
-        samples_summary: af.SamplesSummary,
-        paths: Optional[af.AbstractPaths] = None,
-        samples: Optional[af.SamplesPDF] = None,
-        search_internal: Optional[object] = None,
-        analysis: Optional[AnalysisDataset] = None,
-    ):
-        """
-        After the non-linear search of a fit to a dataset is complete it creates a `Result` object which includes:
-
-        - The samples of the non-linear search (E.g. MCMC chains, nested sampling samples) which are used to compute
-          the maximum likelihood model, posteriors and other properties.
-
-        - The model used to fit the data, which uses the samples to create specific instances of the model (e.g.
-          an instance of the maximum log likelihood model).
-
-        - The non-linear search used to perform the model fit.
-
-        This class contains a number of methods which use the above objects to create the max log likelihood galaxies,
-        `FitIamging`, adapt-galaxy images,etc.
-
-        Parameters
-        ----------
-        samples
-            A PyAutoFit object which contains the samples of the non-linear search, for example the chains of an MCMC
-            run of samples of the nested sampler.
-        model
-            The PyAutoFit model object, which includes model components representing the galaxies that are fitted to
-            the imaging data.
-        analysis
-            The `Analysis` object that was used to perform the model-fit from which this result is inferred.
-        search
-            The non-linear search used to perform this model-fit.
-
-        Returns
-        -------
-        ResultImaging
-            The result of fitting the model to the imaging dataset, via a non-linear search.
-        """
-        super().__init__(
-            samples_summary=samples_summary,
-            paths=paths,
-            samples=samples,
-            search_internal=search_internal,
-        )
-
-        self.analysis = analysis
 
     @property
     def max_log_likelihood_fit(self):
