+++ autogalaxy/imaging/model/result.py
@@ -8,37 +8,6 @@ from autogalaxy.imaging.fit_imaging import FitImaging
 
 
 class ResultImaging(ResultDataset):
-    """
-    After the non-linear search of a fit to an imaging dataset is complete it creates this `ResultImaging`, object
-    which includes:
-
-    - The samples of the non-linear search (E.g. MCMC chains, nested sampling samples) which are used to compute
-    the maximum likelihood model, posteriors and other properties.
-
-    - The model used to fit the data, which uses the samples to create specific instances of the model (e.g.
-    an instance of the maximum log likelihood model).
-
-    - The non-linear search used to perform the model fit.
-
-    This class contains a number of methods which use the above objects to create the max log likelihood galaxies,
-    `FitImaging`, adapt-galaxy images,etc.
-
-    Parameters
-    ----------
-    samples
-        A PyAutoFit object which contains the samples of the non-linear search, for example the chains of an MCMC
-        run of samples of the nested sampler.
-    model
-        The PyAutoFit model object, which includes model components representing the galaxies that are fitted to
-        the imaging data.
-    search
-        The non-linear search used to perform this model-fit.
-
-    Returns
-    -------
-    ResultImaging
-        The result of fitting the model to the imaging dataset, via a non-linear search.
-    """
 
     @property
     def max_log_likelihood_fit(self) -> FitImaging:
