+++ autogalaxy/quantity/model/analysis.py
@@ -1,5 +1,4 @@
 import os
-from typing import Optional
 
 from autoconf.dictable import to_dict
 
@@ -17,6 +16,9 @@ from autogalaxy import exc
 
 
 class AnalysisQuantity(Analysis):
+
+    Result = ResultQuantity
+
     def __init__(
         self,
         dataset: DatasetQuantity,
@@ -161,48 +163,6 @@ class AnalysisQuantity(Analysis):
 
         visualizer = VisualizerQuantity(visualize_path=paths.image_path)
         visualizer.visualize_fit_quantity(fit=fit)
-
-    def make_result(
-        self,
-        samples_summary: af.SamplesSummary,
-        paths: af.AbstractPaths,
-        samples: Optional[af.SamplesPDF] = None,
-        search_internal: Optional[object] = None,
-    ) -> ResultQuantity:
-        """
-        After the non-linear search is complete create its `ResultQuantity`, which includes:
-
-        - The samples of the non-linear search (E.g. MCMC chains, nested sampling samples) which are used to compute
-        the maximum likelihood model, posteriors and other properties.
-
-        - The model used to fit the data, which uses the samples to create specific instances of the model (e.g.
-        an instance of the maximum log likelihood model).
-
-        - The non-linear search used to perform the model fit.
-
-        The `ResultQuantity` object contains a number of methods which use the above objects to create the max
-        log likelihood galaxies, `FitQuantity`,etc.
-
-        Parameters
-        ----------
-        samples
-            A PyAutoFit object which contains the samples of the non-linear search, for example the chains of an MCMC
-            run of samples of the nested sampler.
-        search
-            The non-linear search used to perform this model-fit.
-
-        Returns
-        -------
-        ResultQuantity
-            The result of fitting the model to the imaging dataset, via a non-linear search.
-        """
-        return ResultQuantity(
-            samples_summary=samples_summary,
-            paths=paths,
-            samples=samples,
-            search_internal=search_internal,
-            analysis=self,
-        )
 
     def save_attributes(self, paths: af.DirectoryPaths):
         """
